✨ Added Support for backgroundImage property for every primitive type
