+++ appcues/src/androidTest/java/com/appcues/ui/primitive/VerticalStackComponentTest.kt
@@ -1,4 +1,4 @@
-package com.appcues.ui.component
+package com.appcues.ui.primitive
 
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onRoot
@@ -6,42 +6,33 @@ import com.karumi.shot.ScreenshotTest
 import org.junit.Rule
 import org.junit.Test
 
-class ButtonComponentExtraTest : ScreenshotTest {
+class VerticalStackComponentTest : ScreenshotTest {
 
     @get:Rule
     val composeRule = createComposeRule()
 
     @Test
-    fun testPreviewButtonBorder() {
+    fun testPreviewTestVerticalAlignment() {
         composeRule.setContent {
-            PreviewButtonBorder()
+            PreviewTestVerticalAlignment()
         }
 
         compareScreenshot(composeRule.onRoot())
     }
 
     @Test
-    fun testPreviewButtonComplexContents() {
+    fun testPreviewTestVerticalDefault() {
         composeRule.setContent {
-            PreviewButtonComplexContents()
+            PreviewTestVerticalDefault()
         }
 
         compareScreenshot(composeRule.onRoot())
     }
 
     @Test
-    fun testPreviewButtonDefault() {
+    fun testPreviewTestVerticalLayout() {
         composeRule.setContent {
-            PreviewButtonDefault()
-        }
-
-        compareScreenshot(composeRule.onRoot())
-    }
-
-    @Test
-    fun testPreviewButtonGeneral() {
-        composeRule.setContent {
-            PreviewButtonGeneral()
+            PreviewTestVerticalLayout()
         }
 
         compareScreenshot(composeRule.onRoot())
