+++ appcues/src/main/java/com/appcues/data/mapper/styling/SizeMapper.kt
@@ -1,11 +1,11 @@
 package com.appcues.data.mapper.styling
 
 import com.appcues.data.model.styling.ComponentSize
-import com.appcues.data.remote.response.styling.SizeResponse
+import com.appcues.data.remote.response.styling.StyleSizeResponse
 
 internal class SizeMapper {
 
-    fun map(from: SizeResponse) = ComponentSize(
+    fun map(from: StyleSizeResponse) = ComponentSize(
         width = from.width,
         height = from.height,
     )
