+++ appcues/src/main/java/com/appcues/data/remote/response/styling/StyleBackgroundImageResponse.kt
@@ -1,0 +1,13 @@
+package com.appcues.data.remote.response.styling
+
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+internal data class StyleBackgroundImageResponse(
+    val imageUrl: String,
+    val blurHash: String? = null,
+    val intrinsicSize: StyleSizeResponse? = null,
+    val contentMode: String? = null,
+    val verticalAlignment: String? = null,
+    val horizontalAlignment: String? = null,
+)
