+++ appcues/src/main/java/com/appcues/data/remote/response/styling/StyleSizeResponse.kt
@@ -3,7 +3,7 @@ package com.appcues.data.remote.response.styling
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
-internal data class SizeResponse(
+internal data class StyleSizeResponse(
     val width: Int,
     val height: Int,
 )
