+++ appcues/src/main/java/com/appcues/ui/primitive/HorizontalStackPrimitive.kt
@@ -1,6 +1,5 @@
-package com.appcues.ui.component
+package com.appcues.ui.primitive
 
-import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -8,9 +7,7 @@ import androidx.compose.foundation.layout.BoxScope
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.RowScope
 import androidx.compose.material.ProvideTextStyle
-import androidx.compose.material.ripple.rememberRipple
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
@@ -23,35 +20,19 @@ import com.appcues.data.model.styling.ComponentDistribution
 import com.appcues.data.model.styling.ComponentStyle
 import com.appcues.data.model.styling.ComponentStyle.ComponentHorizontalAlignment
 import com.appcues.data.model.styling.ComponentStyle.ComponentVerticalAlignment
-import com.appcues.ui.LocalAppcuesActionDelegate
-import com.appcues.ui.LocalAppcuesActions
-import com.appcues.ui.arrangement.AppcuesArrangement
-import com.appcues.ui.extensions.Compose
-import com.appcues.ui.extensions.PrimitiveGestureProperties
 import com.appcues.ui.extensions.getTextStyle
 import com.appcues.ui.extensions.getVerticalAlignment
-import com.appcues.ui.extensions.primitiveStyle
-import com.appcues.ui.theme.AppcuesPreview
+import com.appcues.ui.theme.AppcuesPreviewPrimitive
+import com.appcues.ui.utils.AppcuesArrangement
 import java.util.UUID
 
 @Composable
-internal fun HorizontalStackPrimitive.Compose() {
+internal fun HorizontalStackPrimitive.Compose(modifier: Modifier) {
     Row(
-        modifier = Modifier.primitiveStyle(
-            component = this,
-            gestureProperties = PrimitiveGestureProperties(
-                onAction = LocalAppcuesActionDelegate.current.onAction,
-                actions = LocalAppcuesActions.current,
-                interactionSource = remember { MutableInteractionSource() },
-                indication = rememberRipple(),
-                enabled = remember { true },
-            ),
-            isDark = isSystemInDarkTheme()
-        ),
+        modifier = modifier,
         horizontalArrangement = distribution.toHorizontalArrangement(spacing),
         verticalAlignment = style.getVerticalAlignment()
     ) {
-
         ProvideTextStyle(style.getTextStyle(LocalContext.current, isSystemInDarkTheme())) {
             items.forEach {
                 ItemBox(distribution = distribution, style = it.style) {
@@ -137,96 +118,86 @@ private val items = arrayListOf(
 @Composable
 @Preview(name = "stack-horizontal-alignment.json", group = "extra")
 internal fun PreviewTestHorizontalAlignment() {
-    val component = HorizontalStackPrimitive(
-        id = UUID.randomUUID(),
-        style = ComponentStyle(
-            verticalAlignment = ComponentVerticalAlignment.TOP,
-            backgroundColor = ComponentColor(light = 0xFFCDCDFA, dark = 0xFFCDCDFA)
-        ),
-        items = items
-    )
-
-    AppcuesPreview {
-        component.Compose()
+    AppcuesPreviewPrimitive {
+        HorizontalStackPrimitive(
+            id = UUID.randomUUID(),
+            style = ComponentStyle(
+                verticalAlignment = ComponentVerticalAlignment.TOP,
+                backgroundColor = ComponentColor(light = 0xFFCDCDFA, dark = 0xFFCDCDFA)
+            ),
+            items = items
+        )
     }
 }
 
 @Composable
 @Preview(name = "stack-horizontal-default.json", group = "extra")
 internal fun PreviewTestHorizontalDefault() {
-    val component = HorizontalStackPrimitive(
-        id = UUID.randomUUID(),
-        style = ComponentStyle(
-            backgroundColor = ComponentColor(light = 0xFFCDCDFA, dark = 0xFFCDCDFA)
-        ),
-        items = items
-    )
-
-    AppcuesPreview {
-        component.Compose()
+    AppcuesPreviewPrimitive {
+        HorizontalStackPrimitive(
+            id = UUID.randomUUID(),
+            style = ComponentStyle(
+                backgroundColor = ComponentColor(light = 0xFFCDCDFA, dark = 0xFFCDCDFA)
+            ),
+            items = items
+        )
     }
 }
 
 @Composable
 @Preview(name = "stack-horizontal-distribution-equal.json", group = "extra")
 internal fun PreviewTestHorizontalDistributionEqual() {
-    val component = HorizontalStackPrimitive(
-        id = UUID.randomUUID(),
-        distribution = ComponentDistribution.EQUAL,
-        spacing = 10,
-        style = ComponentStyle(
-            width = 400,
-            backgroundColor = ComponentColor(light = 0xFFCDCDFA, dark = 0xFFCDCDFA)
-        ),
-        items = items
-    )
-
-    AppcuesPreview {
-        component.Compose()
+    AppcuesPreviewPrimitive {
+        HorizontalStackPrimitive(
+            id = UUID.randomUUID(),
+            distribution = ComponentDistribution.EQUAL,
+            spacing = 10,
+            style = ComponentStyle(
+                width = 400,
+                backgroundColor = ComponentColor(light = 0xFFCDCDFA, dark = 0xFFCDCDFA)
+            ),
+            items = items
+        )
     }
 }
 
 @Composable
 @Preview(name = "stack-horizontal-distribution-center.json", group = "extra")
 internal fun PreviewTestHorizontalDistributionCenter() {
-    val component = HorizontalStackPrimitive(
-        id = UUID.randomUUID(),
-        distribution = ComponentDistribution.CENTER,
-        spacing = 8,
-        style = ComponentStyle(
-            width = 400,
-            backgroundColor = ComponentColor(light = 0xFFCDCDFA, dark = 0xFFCDCDFA)
-        ),
-        items = items
-    )
-
-    AppcuesPreview {
-        component.Compose()
+    AppcuesPreviewPrimitive {
+        HorizontalStackPrimitive(
+            id = UUID.randomUUID(),
+            distribution = ComponentDistribution.CENTER,
+            spacing = 8,
+            style = ComponentStyle(
+                width = 400,
+                backgroundColor = ComponentColor(light = 0xFFCDCDFA, dark = 0xFFCDCDFA)
+            ),
+            items = items
+        )
     }
 }
 
 @Composable
 @Preview(name = "stack-horizontal-layout.json", group = "extra")
 internal fun PreviewTestHorizontalLayout() {
-    val component = HorizontalStackPrimitive(
-        id = UUID.randomUUID(),
-        spacing = 20,
-        style = ComponentStyle(
-            width = 500,
-            marginTop = 5,
-            marginLeading = 4,
-            marginBottom = 6,
-            marginTrailing = 3,
-            paddingTop = 6,
-            paddingLeading = 7,
-            paddingBottom = 9,
-            paddingTrailing = 10,
-            backgroundColor = ComponentColor(light = 0xFFCDCDFA, dark = 0xFFCDCDFA)
-        ),
-        items = items
-    )
-
-    AppcuesPreview {
-        component.Compose()
+    AppcuesPreviewPrimitive {
+        HorizontalStackPrimitive(
+            id = UUID.randomUUID(),
+            spacing = 20,
+            style = ComponentStyle(
+                width = 500,
+                marginTop = 5,
+                marginLeading = 4,
+                marginBottom = 6,
+                marginTrailing = 3,
+                paddingTop = 6,
+                paddingLeading = 7,
+                paddingBottom = 9,
+                paddingTrailing = 10,
+                backgroundColor = ComponentColor(light = 0xFFCDCDFA, dark = 0xFFCDCDFA)
+            ),
+            items = items
+        )
     }
 }
