+++ appcues/src/test/java/com/appcues/data/mapper/styling/SizeMapperTest.kt
@@ -1,6 +1,6 @@
 package com.appcues.data.mapper.styling
 
-import com.appcues.data.remote.response.styling.SizeResponse
+import com.appcues.data.remote.response.styling.StyleSizeResponse
 import com.google.common.truth.Truth.assertThat
 import org.junit.Test
 
@@ -11,7 +11,7 @@ class SizeMapperTest {
     @Test
     fun `map SHOULD map from SizeResponse to ComponentSize`() {
         // Given
-        val from = SizeResponse(
+        val from = StyleSizeResponse(
             width = 1000,
             height = 2000
         )
